requirements.txt
selenium~=4.8.2
pytest~=7.2.2
allure-pytest~=2.13.1
psycopg2~=2.9.5
pytest-rerunfailures~=11.1.2
pytest-xdist~=3.2.1
pytest-dependency~=0.5.1
pytest-order~=1.1.0
SQLAlchemy~=2.0.6
webdriver-manager~=3.8.5
cx-Oracle~=8.3.0
APScheduler~=3.10.1
pytz~=2023.3
python-decouple~=3.8